Configure project to use gradle instead of ant
The ant build system is great and all (actually, it isn't), but gradle
is way easier to use and understand. I also r4worked the project file
and package structure to better match the maven way, which has
essentially become the de-facto standard for most java projects.

Lastly, I also added some eclipse related files to .gitignore just in
case any team member decides to go against the rules and open the
the project in eclipse.