Change report chart to standard 1 series bar chart.
The old chart was way too confusing when displaying data for several
room categories at once. This new chart is much easier to figure out.

I'm still trying to fix the bug where the data isn't being calculated
properly, but at least we know the chart logic itself works. It's just
the data gathering logic that is borked.

I'm on it.