Implement view guest bill screen logic
All business logic is implemented for this screen now.

Note that the screen is currently populated with test data just for demo
purposes. The actual logic to display the real bill items IS in place and
working, but because the necessary driving logic is incomplete, the test
logic has been left as well. It will be removed in time. Just know that the
current implementation does show the actual bill items properly, in addition
to generating a large number of 'fake' test items.